Touched up Movement
I had found that movement was lagging HARD if the mov stat was over 3.
So I fixed up the recursion a bit, and set it up to spawn moveTiles out
side of the recursive method. Runs quick up to 7 (Didnt check higher
because 7 is the highest mov Im planning for right now. May need to
touch up more in the future)