Summoning Update
Summoning is now a thing. The summoning menu is 90% done (excluding art
assets, of course). Currently, summons always come in to the right of a
summoner. I want to make a bunch of "summon" map tiles as the playet
selectss a character to summon. Then have them select the tile to summon
on to. This will likely be included in the next update.

Attacking is now also a thing. However, currently upon attacking, your
character simply hits EVERYTHING in their range. I am unsure if I want
to keep this as a mechanic, or include some attacking way to choose who
to attack.