Adding a TODO list
I keep forgetting where the project is and what needs to be done and
what Ideas I had. So from here on out this is a TODO list to use.
Everyone feel free to edit, but pls let the rest of us know if you do,
and dont commit anything else if you commit an update to this file. We
want to avoid conflicts as much as possible.