More Optimizations
Ive set the hub as the standard even handler. It will move everything
around, and should be in control over most things. This will keep every
object up to date on what is going on.
The cursor will tell the topbar when it is on a character, so that it
doesnt need to check every frame... I have fixed a bunch of really
stupid Update methods. Wasting way less time.